requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yt-dlp==2024.1.24
redis==5.0.1
aio-pika==13.1.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psutil==5.9.8
httpx==0.25.2
aiofiles==23.2.1
supabase==2.3.0
jinja2==3.1.2
python-multipart==0.0.6
requests>=2.31.0
sqlalchemy==2.0.23
